+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/nik-rev/auto-default/compare/v0.2.4...HEAD
 
+### Fixed
+
+- Fixed parsing logic for types that contain a comma in them (e.g. `HashMap<String, String>`)
+
 ## [v0.2.4] - 2026-02-27
 
 [v0.2.4]: https://github.com/nik-rev/auto-default/compare/v0.2.3...v0.2.4